Fixed borders on options screen
The methodology for including borders separating items on the options
menu was very kludgy. It used a background drawable that was a box with
a black line at the bottom. This caused problems because the background
colour had to be explicitly stated in code, which caused issues with
certain android versions, because not all of them used the same default
background colour. I could've forced the app to use a predefined
background colour, but I thought this would cause more problems than it
would solve. So I did something much more complex, yet paradoxically
simpler, by just adding a line in an ImageView object at the bottom of
each SetSelectionItem, and fiddling around with the Layout containers.
Don't know why I didn't think of it sooner.