Split up the last kana set into two
Because I'm modeling the design of this app with the same progression
used in the game Learn Japanese To Survive, I decided to finally look up
how progression is actually handled in the parts I haven't played yet.
As a result, I learned that the Y and W groups were split up in the game
with the single N consonant paired with the W group, so I modified this
app to reflect that.