Added system to extend anti-repetion function
To stop the quiz from having the same questions appear too frequently, I
added a record of the last few questions that have been asked and a
check to ensure the question hasn't been asked recently, replacing the
previous method which just prevented the same question from appearing
twice in a row. It also includes a check to prevent the limiter from
introducing an infinite loop by having the required number of unique
questions be greater than the number of questions available.