Fixed wonky display
When either of the alternate incorrect answer behaviours are activated
the lblResponse TextView box would jump around in size, causing the
widgets below the box to move up and down. This has been fixed by
setting the box to two lines in size when either of those two options
are active. Also centered the text vertically for aesthetics.