Removed unneccessary delay
I once programmed the app to delay rebuilding the question bank after
someone returned from the option screen by half a second. Because
otherwise the question bank would rebuild before the changes to the
option screen were recorded, making the whole thing useless. But I since
changed something, likely the point when the changes on the option
screen are recorded, that rendered the delay redundant. So I removed it.