Stop screen rotation
Was having a problem with the quiz resetting every time the screen
rotates. Tried finding a method to preserve the score and the current
question between rotations, but it seemed the simpler solution was to
just disable screen rotation and lock the screen in portrait mode,
mainly because the app doesn't appear to be functional in landscape
mode. I also removed some redundant code in MainQuiz.java. There's no
need to initialize the score to zero when the onCreate method does it as
well.