Created new KanaSelectionItem class
Replaced SetSelectionItem with new KanaSelectionItem that's intended to
take greater advantage of the features of Android Studio and the Android
SDK. It uses the XML interface for page layout, rather than coded in the
java files. The new class retains much of the same code from the old
one, copy-pasted over, but reorganized to fit the new structure.

To take full advantage of this, I had to move the constants used in
OptionsControl to strings.xml as string resources, and as a result add
new functions to the OptionsControl class to accept string resources
instead of pure strings.