Improved error handling on QuestionRecord
If the question record was built using a 'capacity' of 1, which should
represent no repetition, it would originally throw an exception, causing
the app to crash. There was no way to run a check before running the
parent constructor, or run it in a try-catch, so I altered the class to
remove an item after every addition, which leaves an empty space
constantly, but is effective in regulating the frequency of repeating
questions, while allowing the ability to disable the whole feature.