Altered buildRow methods
For the sake of efficiency, I altered the priority of the reference
screen's row building methods. The buildRow method now merely places all
the kana in the table with no spacing which is good for the majority of
kana sets. The buildSpecialRow method is used exclusively for sets nine
and ten, since they don't fill out all five vowels and need to be spaced
out.