Consolidated duplicate methods
The entirety of both the HiraganaQuestions and KatakanaQuestions classes
were identical with the exception of the constants defining what each
category meant. I decided to consolidate all the duplicate methods into
a new QuestionManagement class, and rename the two original classes by
dropping off the second half, and managing the interface through a
single constant instance of the class. Functionally they remain a static
classes, with slight changes in how they're interfaced with.