Added header for digraphs with diacritics
Added a special subheader on the reference screen, in the digraphs
section, for digraphs with diacritics. Part of this involved splitting
the getDigraphsReferenceTable method into two methods, one for the
digraphs with diacritics, and one without. Created a new
diacriticDigraphsSelected method. Also made a slight change to how the
reference pages are built.