Reorganized KanaSelectionItem code
Reorganized KanaSelectionItem code to remove redundant storage of data
(storing title and content in both a variable and as a property of the
relevant TextView objects), and have code that was in the updateObject
method split off so it will only be called when the relevant components
are updated rather than call all the code when anything is updated. Also
moved the setOnCheckedChangeListener command for the checkbox to the
init method because I realized it does not need to be updated whenever
the prefId changes, because the handler calls the getPrefId method,
which'll respond to changes to the prefId on the fly. Finally, added a
check to the section for loading object attributes so that the setters
aren't called if there's nothing to set.