Corrected infinite loop
Because of a small oversight, the app would get caught in an infinite
loop when the multiple choice question option was selected and the first
set of both hiragana and katakana were selected. This was because the
check to prevent the infinite loops failed to accommodate for the fact
that in this scenario, there are duplicate answers in the question bank.
So, it now generates a full list of answers in advance, with no
duplicates, for the rest of the code to work with, to stop the infinite
loops from happening.