Fixed glitch involving answer box
If the user selected no kana on the KanaSelection screen, the MainQuiz
screen would disable the text box for submitting answers, and send an
alert telling the user to select a kana set. But once they did that, and
returned to the MainQuiz screen, the text box would not be re-enabled
because the line that was intended to do this job was commented-out in
an old commit (02caec220af86dcb21e879fcf48e4c8ab6500621). This has been
fixed.