Fixed error with multiple choice and retry options
When the 'retry on incorrect' option selected, the multiple choice pad
would not re-enable the multiple choice buttons when the user is asked
to retry. This has been fixed, but I set it up so the incorrect answer
that was just selected remains disabled, because I had trouble changing
the colour back to its default, and it would also be really stupid to
allow the user to resubmit an answer we already said was incorrect.