Stopped code from running redundantly
The slice of code for the multiple choice function that builds a list of
unique answers in the getPossibleAnswers method, used to run every time
the method was called, which was unnecessary since the result would
always be the same every time. Now, it will only run the first time the
method is called after a change to the contents of the question bank.