Built basic database
Designed a database to store the record of the user's answers to the
questions. This will have three primary purposes: to show the user how
well they're doing each day, to select the questions giving preference
to the kana the user is having the most problems with, and to support
the multiple choice option being a bit more clever in the selection of
choices. The database is implemented using the Room persistance library.