Stopped database causing crashes
Database initialization was causing several crashes, starting with the
call to a non-existent application context when it was called during
initial object construction. Then, because it was running on the main
thread, and the Room library won't allow that by default, I had to
specify that I will allow it.