Fixed logging too many incorrects
The retrying question mode had a tendency to record to many questions
asked. No matter how many times the incorrect answer is given on a
retried question, it should count it as only one incorrect answer. Only
a new question should count as an incorrect answer. This has been fixed.