Fixed percentage calculation
Fixed the 'get percentage' methods in LogDao, which were being
calculated as the number of correct answers divided by the number of
incorrect answers, when it should be calculated as the number of correct
answers divided by the total answers. The former could cause
divide-by-zero errors, so it really needed to be fixed.