Adjusted getPossibleAnswers method
Edited the getPossibleAnswers method to operate under the new class
design. Since, unlike a standard array, the keys in a map are not
continuous. So calls to non-existent keys were causing
NullPointerExceptions. The values method and a foreach loop allows the
app to build the fullAnswerList the same way as before without causing
any exceptions