Segregated redundant code
There are major chunks of code in the getPossibleAnswers method that
could be cut out under certain circumstances. For a start, there's no
need to try to fetch the database for the current correct answer, and
add it to the weighted list because it's already selected, so that's
been conditioned out. Secondly, if the fullAnswerList is not larger than
the maximum number of choices we want available, then there's no point
in running the various database calls and randomization methods, since
all the items will be selected regardless. Better to just return the
fullAnswerList instead, with no changes.