Removed doubled addQuestions method
Removed the method used to add two question sets to the KanaQuestionBank
with one method. This method was only there so I could have certain if
statements use a single line of code with no curly braces. It wasn't a
good reason to begin with and doesn't increase efficiency in any way, so
it's been removed.