Weighted answers now grow exponentially
The multiple choice feature, which selects the choices available to the
user based on answers which they got wrong in the past, is now set up so
the probability of certain choices being selected grows exponentially as
they keep selecting them. To prevent integer overflows, I also added a
control with the Math.min function so the growth is capped after they
selected it 24 times. Also added comments explaining the mathematics for
the two weighted selection algorithms.