Streamlined remove method
The remove method has been cleaned up a bit. It'll now check to see if
it is the final item in the list. If not, it will go through and correct
only those ahead of it in the list using the higherKey method. It will
also correct the maxValue property, which is a step I forgot to include
last time.