Replaced ArrayLists with TreeSets
The two ArrayLists in KanaQuestionBank are used to store answers for the
multiple choice section. I realized that both must always be sorted by
gojuon order, and can't have any repeating items. So rather than have
checks in place to ensure that, I realized a Set does this
automatically. So the sets have replaced the lists with little
structural change.