Added debug build type
Added debug build type to build.gradle. It adds a suffix to the
application id, and the version name. This is to prevent the development
version of the app conflicting with the production version on any
devices that are running both, as well as making the two versions
identifiable from the 'About' screen. I'd been doing this before now by
manually editing the application id between releases, and not committing
the change. This edit simplifies that process.