Handled display of insanely high answer counts
New parseCount method contains checks to handle the display of counts over 1000, which would otherwise cause rendering problems. While it's unlikely any user would actually answer that many questions, it's a good idea to handle it regardless. The method is designed to handle any number up to a hundred thousand. There's no need to handle anything higher since, because of the one second delay between answering a question and the next one being displayed, it would be physically impossible for any user to answer more than 86,400 questions in a single day. At least, not without cheating or something.