Redesigned icon
With the help of the Android Asset Studio (romannurik/AndroidAssetStudio), I created a new icon using the Google Noto font, as opposed to the generic Sans font in GIMP. I created a basic image in GIMP with the Noto font, and imported it into the Android Asset Studio to gussy it up and make it fit into the Material Design guidelines. Much prettier.