Added icon credits
Gave credit to the components of the icon that I cannot claim credit for. This includes the font, and the Android Assist Studio which did the final work. I am not crediting GIMP because that's just an image editing software and more of a utility or a tool, it didn't add any artistic elements. However I feel the Android Asset Studio did add some artistic elements to the icon to make it fit into the Material Design guidelines, so that is receiving credit.