Preparing for translators
Adding tags to strings.xml to mark which strings aren't to be translated into any languages. This includes all prefids, which shouldn't be seen by the public, and if they were translated would cause problems if the user switched languages; as well as the words Hiragana and Katakana, which are Japanese, and shouldn't really have a translation in any other language.