Rebuilt DailyLogItem
DailyLogItem is now built similarly to ReferenceCell, as a simple View object built internally with Paint objects and Canvasses, rather than built as a LinearLayout containing TextView objects. This method is more efficient, and will prevent the appearance of the most problematic rendering glitches. With this design, I can remain in control, and am not subject to the word wrapping whims of the TextView. Also, no need to annoyingly tweak the width of the individual elements anymore.