Modified dividing line rendering
Set a colour and appropriate width for the dividing line, and had the line's positioning calculations run at the same time all the other positioning calculations run. Also removed the line's stroke width from the Ypoint calculations for all the text, since I don't know why it was there in the first place.