Added dynamic font feature
Added a new feature to the DailyLogItem class that automatically shrinks the text in the event that it starts to overlap with other text. Only issue so far is that it won't shrink the height of the object accordingly. This feature can be disabled entirely by setting isDynamicSize to false.