Expanded and renamed measure method
Made some changes in the three custom view classes (DailyLogItem, KanaSelectionItem, and ReferenceCell) to the onMeasure and measure methods. Took the block of code with the getMode and getSize method calls, and moved it to the measure method, while consolidating some of the code. Insead of storing the result in variables, just calling the methods when the data is needed. But this change meant the measure method didn't need to recieve the mode and size results, since it would get that from the single measureSpec value. But this created a conflict. There was already a measure method in the parent class that I was unaware of, and changing this method from accepting three integers to accepting two integers meant this method and the parent's method came into conflict. I had to rename this method to calculateSize in order to resolve it.