Changed KanaSelectionItem's parent
Rejiggered KanaSelectionItem to extend the RelativeLayout class, instead of the LinearLayout class. This makes the design slightly less convoluted, since there are no more containers within containers. Most of the work was done in the layout file, but the onMeasure method had to be rewritten to ensure everything was properly calculated, and the call to the parent method is needed to position all the internal objects. Without it, the item appears blank.