Modified compilation scripts
Added google to the list of repositories in the main build.gradle file to allow compiling from the command line to work (mainly for Travis-CI). Then, while I was testing it I got an alert saying that the 'compile' command, and all derived commands, in the app's build.gradle file are deprecated. They now use the term 'implementation.' So I fixed that.