Modified set titles
In an attempt to make translation easier, I split off the word 'set', which appears 14 times in the strings resource file, into a single resource, which is replaced in each of the set_title strings with formatting arguments. And because most of the set titles will no longer need translation, I set them to be untranslatable. Also edited the getSelectionScreen method to utilize this new design.