Rebuilt Question Storage
Rather than have the questions stored as constants in specially-designed classes I moved them all to XML documents. In order to make this work several steps had to be taken. First, the now-anemic Hiragana and Katakana classes have to be initialized in MainQuiz with a context object so it can fetch and parse the XML data, and resolve any resources that are called within. The initialize method just calls a parseXml method within the QuestionManagement class that uses a series of loops, if-else chains, and switch-case blocks to properly parse the XML document and feed the resulting data directly into the singleton objects within the two classes, bypassing the need for any setters. The parseXml method could probably use extensive documentation. The ability for a question to have multiple acceptable answers has not been implemented in the new system. Finally, a single change was made in the app code style guide to ensure that the hiragana and katakana XML files remain readable, without each attribute ending up on its own line.