Created parseXmlValue method
Split off the code from parseXml that pulls data from resources. Had to create two methods, one for when we just need the pure string resource, and another is called when the string needs to be altered with extra information, specifically the word 'set.' The former however, just calls the latter since most of the code is the same, but a conditional operator handles the difference.