Altered parseXml method
Made a few structural changes. Moved a few variables to within the try block, restructured a part of the if-else chain so the start tag check is only run once. Moved the calls to the ArrayList object add methods to the same code block, since all four ArrayList objects should be about the same size at the top dimension, and this rendered a try-catch statement unnecessary, so I removed that. Finally moved a check for a valid currentSetNumber to the top of the if block, where it really should be. Most of this was just window dressing. Should result in no functional changes. But hopefully it will help in simplifying things.