Fixed exception handling
Reorganized methods in XmlParser class to handle exceptions more elegantly. Rather than just use a 'return true' statement to crudely exit the method, a break command is used to exit the loop and naturally proceed onto the end of the method. Exceptions are still thrown, but only if the if-else chain finds the end of document marker unexpectedly. Also altered a few of the thrown messages, one to provide more information, and the other because it was just incorrect.