Removed 'number' attribute
The 'number' attribute in the KanaSet tag I realized was useless since it just incremented like clockwork. So I removed it and allowed the XML parsing methods to set the number naturally, starting from zero and going up. This should only result in problems if the KanaSet elements start getting randomly rearranged, but since the same thing could easily happen to the KanaQuestion elements, I think that's the least of my worries.