Altered no questions check
For some reason the no questions check used to trigger when there was only one question in the KanaQuestionBank. This wouldn't be a problem during normal operation, but I ran into issues while testing an unrelated feature. So now it will only trigger when there really are no questions in the KanaQuestionBank.