Altered loop structure
Modified the structure of the loops in the XmlParser class so they operate without break statements. Instead of having the condition detecting when to exit the loop inside the loop's if-else chain, I moved it to be the loop's primary condition and moved the statements that were in that if block to be just below the loop, so they still run once the loop is done. This changes no functionality, and was just done for the sake of my own sanity.