Moved initialization code to QuestionManagement
There were two sets of code that needed to run every time the app started. The code to initialize the Hiragana and Katakana classes so they can load the information from the XML documents, and the code to compile the KanaQuestionBank. It had to be done for each of the classes. While the code is very simple, and was doing no harm remaining in the MainQuiz class, I decided to move the code to two static method in the QuestionManagement class. So it's only one line of code for each of the processes in MainQuiz, and if any additional categories are added, they will be added to the code in the QuestionManagement class.