Modified scoring system
To give a slight incentive to those using the retry function, retrying a question still nets the user partial points. They get a half point if they get it right after one incorrect. After two incorrect answers, the potential score is now a quarter point. Each incorrect answer cuts the potential points in half. This caps after four incorrect answers, which which brings it down to one-sixteenth of a point. After that, they get no points whatsoever. This still needs to be implemented in the database function. Also renamed the isNewQuestion variable to isGetNewQuestion for clarity.