Storage updated for new scoring system
New scoring system is now implemented in the logs. Correct_answers are now stored as floating point numbers, and the incorrect_answers field is now the total_answers because the alternative was storing it as a float and doing some wonky math so they still produce the total when added together. Implemented a migration routine for anyone upgrading to the new version. The DailyLogItem displays the new format correctly, displaying the correct answer count with an appropriate number of decimal places. Finally, added a couple things to the gradle script: A block that allows the database schema archive to display in the IDE, and following up from an earlier commit (3bf2e855278e8deb0239e89b9cebb54e84248120) I updated two uses of the 'compile' command which I missed the last time, since that command has been deprecated.