Fixed error in compiling weightedAnswerList
The database calls that were used to compile the weighted answer lists were using the wrong data. Rather than looking up the kana, like it was supposed to, along with other answers; it was looking up the correct answer along with all other answers, which would always turn up no results because the database doesn't store the information that way. Obviously, I fixed it.