Now caching answer lists
So we don't have to call the database a dozen times for every occasion the question changes, I decided to start caching the lists compiled from the database. The one downside to this is that any time an incorrect answer is given, this list won't update automatically. This is probably something that should be implemented... later.