Added and altered WeightedListTest
Added a new component to WeightedListTest. It now checks that trying to get an index greater than or equal to the maxValue will result in an exception. Also trimmed a bit of the fat, for the 'sum' variable was rendered redundant because it's value was equal to the 'i' incrementer by the time it was actually used.