Changed MAX_RETRIES
Changed the MAX_RETRIES value to three, which means a retried answer will score at minimum one-eighth of a point, or no score at all. This was done for two reasons: because awarding a measly one-sixteenth of a point seems like it's taking the piss, and the display of a one-sixteenth fraction is really awful.