Modified sixteenth rendering
Altered the rendering of a sixteenth fraction to not use the fromHtml method. Since I discovered how fractions are rendered using the unicode fraction slash symbol, I realized it wasn't needed. As part of this I needed to change the zero-width space from being defined by an HTML entity to a unicode codepoint.