KanaSelectionItem now implements Checkable
Altered the KanaSelectionItem class to implement the Checkable interface. No specific reason I did this yet, other than as part of a vein attempt to fix a glitch that popped up when I tried to add a third tab to the KanaSelection screen. It didn't work, but I'm keeping this change.