Added system attribute to AltAnswer
AltAnswer now requires the 'system' attribute to identify which romanization system the answer is sourced from. The implementation in the XSD document is quite kludgy, since I have to list not only the three systems, but the various combinations of the three systems in every possible order so all our bases are covered. There's also the UNKNOWN option, which is currently used to refer to the answers I sourced from TextFugu, and cannot find a second source for.