Implemented recent XML changes into parser
XmlParser now reads the 'system' attribute in the AltAnswer element, and stores this information in the KanaQuestion objects, assigning it as the key to the alt answer, which is now stored in a map. Also, since the currently unused WordQuestion element no longer allows AltAnswers, they've been removed from the parser and the WordQuestion class.