Dev build icon differentiated
In order to make development just a bit easier, I added a new string resource file, that's only used in development builds, that changes the app name so it's visually identifiable as different from the release build. This is mainly an extension to an earlier commit (a1d1e9bf7d515adbff2a4479d6ce7833a7e61241) that just added a suffix to the version name and the app id to prevent conflicts. This prevents visual conflicts as well.