Altered GojuonOrderTest for flexibility
Redesigned GojuonOrderTest to run four separate tests, one for each of the romanization systems. This is mainly so I can test each of the systems fully, without having to test scenarios that'll never happen, like the possibility of both 'si' and 'shi,' which are from two different systems, being compared together. This'll give me more flexibility in the redesign of the GojuonOrder class.