Rejiggered how checkAnswer is called
Rather than call the checkAnswer method by casting the context and calling the public method from that, I decided to change it to work through specially designed event handlers. This allows me to set the method to be private, and it makes more sense because using the getContext method is a bit ambiguous.