Redesigned DynamicSize operation
The DynamicSize option in DailyLogItem no longer resizes the date strings. This is because the date is the smallest text in the object, and making it too small could make it unreadable. Also, it means there's no need to try to resize the object to remove unnecessary whitespace, because the date would keep it at the current size and fill the whitespace.