Modified choice selection weight
Redesigned the overflow prevention methodology for the weight calculation in weightedAnswerList so it doesn't just cap at 24 creating a flat top of probability. Instead, it ensures the values all proportionally fit within the range of zero to 24 through various calculations before running the main calculation and adding them to the list.