Altered overflow prevention to be more dynamic
Instead of using 24 as the maximum value for the incorrect answer count, a number which is based on the fact that there are at most 108 possible values that could come up for the answer list, I decided to rewrite it so it chooses a number based on the number of actual possible answers that the system will consider and weight in the list, using the same calculation I to get 24. It also caches the number so the calculation only has to run once. I also altered a comment to fit the line wrapping settings.