Set weightedAnswerListCache to reset on change
The weightedAnswerListCache will now, much like the previousQuestions list and the maxAnswerWeight, reset when there's any change to the overall list. Specifically whenever questions are added. This would only be a problem if there are more questions added after the quiz starts, which the app doesn't do. But in terms of object-oriented design, I thought it was probably a good idea.