Altered second getRandom method
The second getRandom method now fetches unique random items without having to loop through redundantly. Because I wanted to avoid a scenario where the weighting of one item is significantly higher than any other, and as a result the loop runs thousands of times while fetching the same item, which would be ignored because it's already been fetched. To do this I first clone the current WeightedList object, and remove each item that's been randomly fetched so it's not fetched again. Simple and effective. Had to implement a clone method. Also wrote a getRandomKey method to avoid repeating code.