Moved next statements
All the calls of the next method in the very long switch-case block have been moved to the called methods, getSubKey and getAltIKey. The few cases where the next method is not supposed to be called, the previous method is called instead. It's better to design the methods in mind with what is normal, than what is the exception.