Gojuon sort now considers the end of a word.
Added way to check for the end of a string, which is designed to place the string before another string that is identical up to that point. Spaces are also counted since they would act as the end of a word. Also, in order for the check to work, I had to change how the iterative loop operated. Now it only exits when it finds both strings have reached their end.