Moved 'n' consonant handling
Moved the code that deals with the 'n' consonant to the main getSortId method. This consolidates repeating code, since the handling of the 'n' consonant, and the handling of identifying the non-romanji sequences, are quite similar; and simplifies the handling of the 'n' kana group. This will certainly help in simplifying the overall sort method.