Added modules.xml to .gitignore
I discovered that the newest versions of Android Studio automatically build the modules.xml file on gradle sync, so there's no need to keep it in the repository. Especially since the iml files it refers to are also absent. This is also because it's listed in a new project's default gitignore file. So it's been removed, and added to this project's gitignore file.