Fixed crashes caused by Android's memory-saving features
Android's built-in memory-saving features were causing the app to crash if it was restarted while on any screen other than the MainQuiz and AboutScreen activities. This was mainly because the OptionControl, QuestionManagement, and LogDatabase manager classes needed to be initialized before use, and only the MainQuiz class was actually doing that, because I presumed that would be the first activity loaded in all cases. Unfortunately, the memory-saving features were clearing out the memory, including the manager classes that needed to be initialized, but remembering that it was not on the MainQuiz screen, and immediately trying to use those classes, causing NullPointerExceptions. I fixed this by first splitting off the code to run the initialization methods away from the MainQuiz class, and placing it in a new AppTools class, and calling that in the onCreate method of every activity that needed these manager classes.