Updated to Java 8
Modified build script to use Java 8. This changes nothing with regards to compatibility. The app will still run on Android versions as early as 4.2 Jelly Bean (API 17). But certain Java 8 exclusive features will need to be worked around for any versions earlier than 7.0 Nougat (API 24). This is not an issue yet.