Added caches to gitignore
After recently upgrading to Android Studio 3.0, I kept seeing a new file pop up called 'build_file_checksums.ser'. Since this is in a 'cache' folder, and I didn't write it, and it's probably rebuilt in response to any change in the program, I decided to add the entire cache folder to the gitignore file. The alternative of adding it to the repository would be too problematic, and I see no reason to save it, unlike the archived database schemas which need to be recorded for migration testing.