Upgraded Gradle
Upgraded to a newer version of Gradle, 4.5.1. The latest version is 4.10, but the latest Android Gradle plugin has problems with Gradle 4.6 and later. A feature called 'configuration on demand,' is simply not supported on the newer Gradle versions. While I might be able to disable 'configuration on demand' so I can use newer Gradle versions; I decided to stick with the simpler solution, which is to stay just behind Gradle 4.6... for now.