Added rotation support
The quiz can now be played with the device in landscape mode. This was done by using an alternate layout file for when the device is horizontal. I also edited the manifest to allow almost all screens to be viewed in landscape mode as well. However the Kana Selection screen is still only available in portrait mode, because I haven't yet found a way to stop the app from clearing all checkboxes when the device is rotated. The Reference screen is also available in landscape mode, but I feel it has too much whitespace, so I may have to do something about that, at some point in future.