Some code cleanup
Did several things to cleanup some code. In AppTools, I removed a redundant declaration. In WeightedListTest, I altered an internal class to be static, which should save memory. And in the AboutScreen class I removed the final declaration from a String variable, which I'm sure was necessary at some point, but doesn't appear to be anymore.