Removed redundant conditionals
There were two conditional statements in the isDiacritic method that would always return true because they were dealt with at an earlier point in the if-else chain. So they've been removed. I did keep them recorded as commented notes on the right however, just as a reminder that the functionality is dealt with.