Moved parser object to local scope
I originally wanted the parser object to be stored as a class level variable, mainly because I planned to have the question xml files parsed on the fly, as the app needed the information; rather than what I'm doing now, which is pull all the information from the file, and store it in individual storage arrays. But as I looked into it, moving through the xml, using the parser, in an arbitrary fashion, back and forth as the need arises, seems to be impossible, as there is no 'previous' method, or any way to return to an arbitrary index point or line number in the file. So I moved the parser object from being a class level variable, to a local scope within the parseXml method. After all, there's no reason at all to store it for continued use, since it's useless once it gets to the end of the file.