Reorganized QuestionManagement constructor
Moved the code from the parseXml method to the constructor, which is its only use. Also moved the looping logic within that block of code to the XmlParser class because really, that's where it should be as it's a crucial part of the parsing logic. Finally edited the catch block to actually catch and throw a RuntimeException. Because if there is a problem parsing the XML files, that can't be ignored, and it'll mean the entire app is useless. Also removed the pointless isInitialized property.