Converted statement to use lambda notation
The code that runs when a user selects an answer from the MultipleChoicePad is no longer in a private subclass called AnswerListener, but has been moved to a special method, which the button's onClickListener is directed to by a lambda statement. This is simpler, and I assume it'll be more efficient.