Replaced all uses of java.util.Date
Decided to stop using the java.util.Date class, with the intention of switching to the java.time package included in Java 8. Unfortunately, in order to maintain compatibility with any version of Android pre-Oreo (version 8.0, API 26), which is currently 85 per cent of Android users, I found this was not an option. Thankfully, I ran across the Joda-Time packages, which is an improved date and time handling system quite similar to the java.time package in Java 8, but compatible with Java versions as early as Java 5. So I decided to remove java.util.Date, and take full advantage of Joda-Time. At a later date I might be able to drop in the java.time package with minimal changes, but for now I see this as a significant improvement, since java.util.Date and its associated detritus was the ultimate in shambolic design.