Fix build issues
Proguard was having issues with joda-time. I'm not exactly sure why, but I found the solution was adding another package called joda-convert. Unfortunately, this was causing problems with the Java 8 compiler's desugar step. The solution I found was to use the 'classic' variant of joda-convert, but I'm a bit bothered with this solution. However until the wider issue (something about module-info.class) is fixed by either the Gradle or Android teams, or whoever is responsible, there's not much I can do about it.