Significantly improved multiple choice selection
The multiple choice answer option is now enhanced by a category system. Diacritic questions are now four times more likely to feature multiple choice answers that are for diacritics. Same goes for digraphs, questions that are both digraphs and diacritics, and questions that are neither. This should mean that that the choices that appear fit better with the question. I may still think of ways to enhance this further, but for now, this is a major improvement.