Built test for xmlParser
Started writing a test for the XML parser. As part of this, I also wrote a specialized XML document to run the test on, since it's merely meant to test the parser, and if the hiragana and katakana documents change, I don't see the need to alter the test to fit it, since that is not what's being tested. I also had to make a slight alteration to the parameterized fetchRomanji method in the KanaQuestion class, so I can properly examine the KanaQuestion objects returned by the parser without it throwing a NullPointerException. Also, the test itself still needs to examine the third set in the XML document, so I'm classifying it as incomplete.