Changed altRomanji to default to null
As a minor memory-saving thing, I decided to set the altRomanji property in KanaQuestion to default to null when empty, so we're not wasting memory space storing an empty TreeMap object for no reason. This required me adding an extra check in each of the three methods that actually access the altRomanji list, which may waste cycles, but I think the memory saved is a fair trade off.